refactor: break up giant file into less files
