+++ tonal_hortator/core/deduplication.py
@@ -1,0 +1,229 @@
+#!/usr/bin/env python3
+"""
+Deduplication utilities for playlist generation
+Handles multiple deduplication strategies for track lists
+"""
+
+import logging
+import re
+from typing import Any, Dict, List, Optional
+
+logger = logging.getLogger(__name__)
+
+
+class TrackDeduplicator:
+    """Handle track deduplication using multiple strategies"""
+
+    def __init__(self) -> None:
+        self.suffixes_to_remove = [
+            " (remix)",
+            " (remastered)",
+            " (live)",
+            " (acoustic)",
+            " (radio edit)",
+            " (extended)",
+            " (clean)",
+            " (explicit)",
+            " (original mix)",
+            " (club mix)",
+            " (dub mix)",
+            " - remix",
+            " - remastered",
+            " - live",
+            " - acoustic",
+        ]
+
+    def deduplicate_tracks(
+        self, tracks: List[Dict[str, Any]], min_similarity: float, max_tracks: int
+    ) -> List[Dict[str, Any]]:
+        """Apply comprehensive deduplication to track list"""
+        # Filter by similarity threshold
+        filtered = [
+            track
+            for track in tracks
+            if track.get("similarity_score", 0) >= min_similarity
+        ]
+
+        logger.info(
+            f"🔍 Starting deduplication on {len(filtered)} tracks (similarity ≥ {min_similarity})"
+        )
+
+        # Apply multiple deduplication strategies
+        location_deduplicated = self._deduplicate_by_location(filtered)
+        combination_deduplicated = self._deduplicate_by_title_artist(
+            location_deduplicated
+        )
+        id_deduplicated = self._deduplicate_by_track_id(combination_deduplicated)
+        smart_deduplicated = self._smart_title_deduplication(id_deduplicated)
+
+        logger.info(
+            f"✅ Final deduplication summary: {len(tracks)} → {len(smart_deduplicated)} tracks"
+        )
+
+        return smart_deduplicated
+
+    def _deduplicate_by_location(
+        self, tracks: List[Dict[str, Any]]
+    ) -> List[Dict[str, Any]]:
+        """Deduplicate by file location (normalized path)"""
+        seen_locations = set()
+        location_deduplicated = []
+
+        for track in tracks:
+            location = self._normalize_file_location(track.get("location", ""))
+            if location and location not in seen_locations:
+                seen_locations.add(location)
+                location_deduplicated.append(track)
+            elif not location:  # Include tracks without location
+                location_deduplicated.append(track)
+
+        logger.info(
+            f"📍 Location deduplication: {len(tracks)} → {len(location_deduplicated)} tracks"
+        )
+        return location_deduplicated
+
+    def _deduplicate_by_title_artist(
+        self, tracks: List[Dict[str, Any]]
+    ) -> List[Dict[str, Any]]:
+        """Deduplicate by title/artist combination"""
+        seen_combinations = set()
+        combination_deduplicated = []
+
+        for track in tracks:
+            title = track.get("name", "").strip().lower()
+            artist = track.get("artist", "").strip().lower()
+            combination = f"{title}|{artist}"
+
+            if combination and combination not in seen_combinations:
+                seen_combinations.add(combination)
+                combination_deduplicated.append(track)
+            elif not combination:  # Include tracks without title/artist
+                combination_deduplicated.append(track)
+
+        logger.info(
+            f"🎵 Title/Artist deduplication: {len(tracks)} → {len(combination_deduplicated)} tracks"
+        )
+        return combination_deduplicated
+
+    def _deduplicate_by_track_id(
+        self, tracks: List[Dict[str, Any]]
+    ) -> List[Dict[str, Any]]:
+        """Deduplicate by track ID (shouldn't happen, but just in case)"""
+        seen_ids = set()
+        final_deduplicated = []
+
+        for track in tracks:
+            track_id = track.get("id")
+            if track_id not in seen_ids:
+                seen_ids.add(track_id)
+                final_deduplicated.append(track)
+
+        logger.info(
+            f"🆔 Track ID deduplication: {len(tracks)} → {len(final_deduplicated)} tracks"
+        )
+        return final_deduplicated
+
+    def _smart_title_deduplication(
+        self, tracks: List[Dict[str, Any]]
+    ) -> List[Dict[str, Any]]:
+        """Smart deduplication for titles with slight variations (e.g., 'Song (Remix)' vs 'Song')"""
+        if len(tracks) <= 1:
+            return tracks
+
+        artist_groups = self._group_tracks_by_artist(tracks)
+        deduplicated = []
+
+        for artist_tracks in artist_groups.values():
+            deduplicated.extend(self._process_artist_tracks(artist_tracks))
+
+        logger.info(
+            f"🧠 Smart title deduplication: {len(tracks)} → {len(deduplicated)} tracks"
+        )
+        return deduplicated
+
+    def _group_tracks_by_artist(self, tracks: List[Dict[str, Any]]) -> dict[str, list]:
+        """Group tracks by artist for deduplication"""
+        artist_groups: dict[str, list] = {}
+        for track in tracks:
+            artist = track.get("artist", "").strip().lower()
+            if artist:
+                if artist not in artist_groups:
+                    artist_groups[artist] = []
+                artist_groups[artist].append(track)
+        return artist_groups
+
+    def _process_artist_tracks(
+        self, artist_tracks: List[Dict[str, Any]]
+    ) -> List[Dict[str, Any]]:
+        """Process tracks for a single artist to remove duplicates"""
+        if len(artist_tracks) == 1:
+            return [artist_tracks[0]]
+
+        deduplicated = []
+        processed_titles = set()
+
+        for track in artist_tracks:
+            title = track.get("name", "").strip().lower()
+            base_title = self._extract_base_title(title)
+
+            if base_title and base_title not in processed_titles:
+                processed_titles.add(base_title)
+                best_track = self._find_best_track_for_base_title(
+                    artist_tracks, base_title
+                )
+                if best_track:
+                    deduplicated.append(best_track)
+            elif not base_title:
+                deduplicated.append(track)
+
+        return deduplicated
+
+    def _find_best_track_for_base_title(
+        self, artist_tracks: List[Dict[str, Any]], base_title: str
+    ) -> Optional[Dict[str, Any]]:
+        """Find the track with the highest similarity score for a given base title"""
+        best_track = None
+        best_score = -1
+        for track in artist_tracks:
+            title = track.get("name", "").strip().lower()
+            track_base = self._extract_base_title(title)
+            if track_base == base_title:
+                score = track.get("similarity_score", 0)
+                if score > best_score:
+                    best_score = score
+                    best_track = track
+        return best_track
+
+    def _extract_base_title(self, title: str) -> str:
+        """Extract base title by removing common suffixes and variations"""
+        if not title:
+            return ""
+
+        # Remove common suffixes in parentheses
+        base = re.sub(r"\s*\([^)]*\)\s*$", "", title)
+
+        # Remove common suffixes
+        for suffix in self.suffixes_to_remove:
+            if base.lower().endswith(suffix.lower()):
+                base = base[: -len(suffix)]
+                break
+
+        return base.strip()
+
+    def _normalize_file_location(self, location: str) -> str:
+        """Normalize file location for deduplication and Apple Music compatibility"""
+        if not location:
+            return ""
+
+        # Convert to lowercase and normalize separators
+        normalized = location.lower().replace("\\", "/")
+
+        # Remove common OS-specific prefixes with username
+        # Match /users/<username>/, /home/<username>/, c:/users/<username>/, etc.
+        match = re.match(
+            r"^(?:/users/|/home/|c:/users/|d:/users/|e:/users/)([^/]+)/(.+)$",
+            normalized,
+        )
+        if match:
+            return match.group(2)
+        return normalized
