+++ tonal_hortator/core/playlist_generator.py
@@ -5,14 +5,13 @@ This script creates playlists based on semantic search queries
 """
 
 import logging
-import os
-import random
-import re
 import time
-from datetime import datetime
 from typing import Any, Dict, List, Optional
 
+from tonal_hortator.core.artist_distributor import ArtistDistributor
+from tonal_hortator.core.deduplication import TrackDeduplicator
 from tonal_hortator.core.embeddings import OllamaEmbeddingService
+from tonal_hortator.core.query_processor import QueryProcessor
 from tonal_hortator.core.track_embedder import LocalTrackEmbedder
 
 # Configure logging
@@ -43,6 +42,11 @@ class LocalPlaylistGenerator:
             db_path, embedding_service=self.embedding_service
         )
 
+        # Initialize utility classes
+        self.query_processor = QueryProcessor()
+        self.deduplicator = TrackDeduplicator()
+        self.artist_distributor = ArtistDistributor()
+
     def generate_playlist(
         self, query: str, max_tracks: int = 20, min_similarity: float = 0.3
     ) -> List[Dict[str, Any]]:
@@ -61,70 +65,103 @@ class LocalPlaylistGenerator:
             logger.info(f"🎵 Generating playlist for query: '{query}'")
             start_time = time.time()
 
-            # 1. Extract track count from query
-            extracted_count = self._extract_track_count(query)
-            if extracted_count:
-                max_tracks = extracted_count
-                logger.info(
-                    f"🔢 Found track count in query, setting max_tracks to {max_tracks}"
+            # 1. Extract parameters from query
+            max_tracks = self._extract_playlist_parameters(query, max_tracks)
+
+            # 2. Get embeddings and track data
+            embeddings, track_data = self._get_embeddings_for_search()
+            if not embeddings:
+                return []
+
+            # 3. Perform similarity search
+            if track_data:
+                results = self.embedding_service.similarity_search(
+                    query, embeddings, track_data, top_k=max_tracks * 5
                 )
+            else:
+                results = []
 
-            # 2. Get embeddings from database
-            embeddings, track_data = self.track_embedder.get_all_embeddings()
+            # 4. Filter by artist if specified
+            results = self._filter_by_artist(query, results)
 
-            if not embeddings or not track_data:
-                logger.warning(
-                    "No embeddings found. Running embedding process first..."
-                )
-                self.track_embedder.embed_all_tracks()
-                embeddings, track_data = self.track_embedder.get_all_embeddings()
-
-                if not embeddings:
-                    logger.error(
-                        "❌ Still no embeddings available after embedding process"
-                    )
-                    return []
-
-            logger.info(f"📊 Using {len(embeddings)} track embeddings for search")
-
-            # Perform similarity search
-            results = self.embedding_service.similarity_search(
-                query, embeddings, track_data, top_k=max_tracks * 5
-            )
-
-            # 3. Filter by artist if detected
-            artist = self._extract_artist_from_query(query)
-            if artist:
-                logger.info(f"🎤 Found artist in query: '{artist}'. Filtering results.")
-                results = [
-                    track
-                    for track in results
-                    if artist.lower() in track.get("artist", "").lower()
-                ]
-
-            # 4. Apply genre filtering and boosting
-            results = self._apply_genre_filtering(query, results)
-
-            # 5. Filter by similarity threshold and deduplicate
-            filtered_results = self._filter_and_deduplicate_results(
-                results, min_similarity, max_tracks
-            )
-
-            # 6. Apply artist randomization
-            filtered_results = self._apply_artist_randomization(
-                filtered_results, max_tracks, self._is_vague_query(query)
+            # 5. Process and refine the playlist
+            final_playlist = self._process_and_refine_playlist(
+                query, results, min_similarity, max_tracks
             )
 
             generation_time = time.time() - start_time
             logger.info(
-                f"✅ Generated playlist with {len(filtered_results)} tracks in {generation_time:.2f}s"
+                f"✅ Generated playlist with {len(final_playlist)} tracks in {generation_time:.2f}s"
             )
 
-            return filtered_results
+            return final_playlist
 
         except Exception as e:
             logger.error(f"❌ Error generating playlist: {e}")
             raise
+
+    def _extract_playlist_parameters(self, query: str, default_max_tracks: int) -> int:
+        """Extract playlist parameters from the query"""
+        extracted_count = self.query_processor.extract_track_count(query)
+        if extracted_count:
+            logger.info(
+                f"🔢 Found track count in query, setting max_tracks to {extracted_count}"
+            )
+            return extracted_count
+        return default_max_tracks
+
+    def _get_embeddings_for_search(
+        self,
+    ) -> tuple[Optional[List[Any]], Optional[List[Dict[str, Any]]]]:
+        """Get embeddings from the database, generating them if necessary"""
+        embeddings, track_data = self.track_embedder.get_all_embeddings()
+        if not embeddings or not track_data:
+            logger.warning("No embeddings found. Running embedding process first...")
+            self.track_embedder.embed_all_tracks()
+            embeddings, track_data = self.track_embedder.get_all_embeddings()
+            if not embeddings:
+                logger.error("❌ Still no embeddings available after embedding process")
+                return None, None
+        logger.info(f"📊 Using {len(embeddings)} track embeddings for search")
+        return embeddings, track_data
+
+    def _filter_by_artist(
+        self, query: str, results: List[Dict[str, Any]]
+    ) -> List[Dict[str, Any]]:
+        """Filter results by artist if specified in the query"""
+        artist = self.query_processor.extract_artist_from_query(query)
+        if artist:
+            logger.info(f"🎤 Found artist in query: '{artist}'. Filtering results.")
+            return [
+                track
+                for track in results
+                if artist.lower() in track.get("artist", "").lower()
+            ]
+        return results
+
+    def _process_and_refine_playlist(
+        self,
+        query: str,
+        results: List[Dict[str, Any]],
+        min_similarity: float,
+        max_tracks: int,
+    ) -> List[Dict[str, Any]]:
+        """Apply genre filtering, deduplication, and artist randomization"""
+        # Apply genre filtering and boosting
+        results = self._apply_genre_filtering(query, results)
+
+        # Filter by similarity threshold and deduplicate
+        results = self.deduplicator.deduplicate_tracks(
+            results, min_similarity, max_tracks
+        )
+
+        # Apply artist randomization
+        is_vague = self.query_processor.is_vague_query(query)
+        results = self.artist_distributor.apply_artist_randomization(
+            results, max_tracks, is_vague
+        )
+
+        return results
 
     def _apply_genre_filtering(
         self, query: str, results: List[Dict[str, Any]]
@@ -144,7 +181,7 @@ class LocalPlaylistGenerator:
 
         # Extract potential genres from query
         query_lower = query.lower()
-        genre_keywords = self._extract_genre_keywords(query_lower)
+        genre_keywords = self.query_processor.extract_genre_keywords(query_lower)
 
         if not genre_keywords:
             logger.info(
@@ -189,793 +226,3 @@ class LocalPlaylistGenerator:
         )
 
         return combined_results
-
-    def _extract_genre_keywords(self, query: str) -> List[str]:
-        """
-        Extract potential genre keywords from a query
-
-        Args:
-            query: Lowercase search query
-
-        Returns:
-            List of detected genre keywords
-        """
-        # Common music genres and their variations
-        genre_mapping = {
-            "jazz": ["jazz", "bebop", "swing", "fusion", "smooth jazz", "acid jazz"],
-            "rock": [
-                "rock",
-                "hard rock",
-                "soft rock",
-                "classic rock",
-                "punk rock",
-                "indie rock",
-            ],
-            "pop": ["pop", "pop rock", "synth pop", "indie pop"],
-            "hip hop": ["hip hop", "rap", "trap", "r&b", "soul"],
-            "electronic": [
-                "electronic",
-                "edm",
-                "techno",
-                "house",
-                "trance",
-                "ambient",
-                "dubstep",
-            ],
-            "country": ["country", "folk", "bluegrass", "americana"],
-            "classical": ["classical", "orchestral", "symphony", "chamber"],
-            "blues": ["blues", "delta blues", "electric blues"],
-            "reggae": ["reggae", "dub", "ska"],
-            "metal": ["metal", "heavy metal", "death metal", "black metal"],
-            "funk": ["funk", "disco", "motown"],
-            "latin": ["latin", "salsa", "bossa nova", "tango"],
-            "world": ["world", "ethnic", "traditional"],
-            "soundtrack": ["soundtrack", "score", "film music"],
-            "christian": ["christian", "gospel", "worship"],
-            "children": ["children", "kids", "nursery"],
-            "comedy": ["comedy", "humor", "parody"],
-            "holiday": ["holiday", "christmas", "xmas", "seasonal"],
-        }
-
-        detected_genres = []
-
-        for genre, keywords in genre_mapping.items():
-            if any(keyword in query for keyword in keywords):
-                detected_genres.append(genre)
-
-        return detected_genres
-
-    def _extract_artist_from_query(self, query: str) -> Optional[str]:
-        """
-        Extract artist from a query using patterns like "by [artist]" or "[artist] songs"
-        """
-        # Patterns to detect artist in query
-        patterns = [
-            r"by\s+(.+)",  # "songs by [artist]"
-            r"(.+?)\s+songs",  # "[artist] songs"
-            r"(.+?)\s+radio",  # "[artist] radio"
-        ]
-
-        query_lower = query.lower()
-
-        for pattern in patterns:
-            match = re.search(pattern, query_lower)
-            if match:
-                artist = match.group(1).strip()
-                # Capitalize the artist name for better matching
-                return artist.title()
-
-        return None
-
-    def _filter_and_deduplicate_results(
-        self, results: List[Dict[str, Any]], min_similarity: float, max_tracks: int
-    ) -> List[Dict[str, Any]]:
-        """Filter results by similarity and remove duplicates using multiple strategies"""
-        # Filter by similarity threshold
-        filtered = [
-            track
-            for track in results
-            if track.get("similarity_score", 0) >= min_similarity
-        ]
-
-        logger.info(
-            f"🔍 Starting deduplication on {len(filtered)} tracks (similarity ≥ {min_similarity})"
-        )
-
-        # Strategy 1: Deduplicate by file location (normalized path)
-        seen_locations = set()
-        location_deduplicated = []
-
-        for track in filtered:
-            location = self._normalize_file_location(track.get("location", ""))
-            if location and location not in seen_locations:
-                seen_locations.add(location)
-                location_deduplicated.append(track)
-            elif not location:  # Include tracks without location
-                location_deduplicated.append(track)
-
-        logger.info(
-            f"📍 Location deduplication: {len(filtered)} → {len(location_deduplicated)} tracks"
-        )
-
-        # Strategy 2: Deduplicate by title/artist combination
-        seen_combinations = set()
-        combination_deduplicated = []
-
-        for track in location_deduplicated:
-            title = track.get("name", "").strip().lower()
-            artist = track.get("artist", "").strip().lower()
-            combination = f"{title}|{artist}"
-
-            if combination and combination not in seen_combinations:
-                seen_combinations.add(combination)
-                combination_deduplicated.append(track)
-            elif not combination:  # Include tracks without title/artist
-                combination_deduplicated.append(track)
-
-        logger.info(
-            f"🎵 Title/Artist deduplication: {len(location_deduplicated)} → {len(combination_deduplicated)} tracks"
-        )
-
-        # Strategy 3: Deduplicate by track ID (shouldn't happen, but just in case)
-        seen_ids = set()
-        final_deduplicated = []
-
-        for track in combination_deduplicated:
-            track_id = track.get("id")
-            if track_id not in seen_ids:
-                seen_ids.add(track_id)
-                final_deduplicated.append(track)
-
-        logger.info(
-            f"🆔 Track ID deduplication: {len(combination_deduplicated)} → {len(final_deduplicated)} tracks"
-        )
-
-        # Strategy 4: Smart deduplication for similar titles with slight variations
-        smart_deduplicated = self._smart_title_deduplication(final_deduplicated)
-
-        logger.info(
-            f"🧠 Smart title deduplication: {len(final_deduplicated)} → {len(smart_deduplicated)} tracks"
-        )
-
-        # Strategy 5: Enforce artist diversity (NEW)
-        # Calculate max tracks per artist based on playlist size
-        max_tracks_per_artist = max(1, min(3, max_tracks // 5))  # 1-3 tracks per artist
-        diverse_tracks = self._enforce_artist_diversity(
-            smart_deduplicated, max_tracks, max_tracks_per_artist
-        )
-
-        logger.info(
-            f"✅ Final deduplication summary: {len(results)} → {len(diverse_tracks)} tracks"
-        )
-
-        return diverse_tracks
-
-    def _group_tracks_by_artist(self, tracks: List[Dict[str, Any]]) -> dict[str, list]:
-        """Group tracks by artist for deduplication"""
-        artist_groups: dict[str, list] = {}
-        for track in tracks:
-            artist = track.get("artist", "").strip().lower()
-            if artist:
-                if artist not in artist_groups:
-                    artist_groups[artist] = []
-                artist_groups[artist].append(track)
-        return artist_groups
-
-    def _find_best_track_for_base_title(
-        self, artist_tracks: List[Dict[str, Any]], base_title: str
-    ) -> Optional[Dict[str, Any]]:
-        """Find the track with the highest similarity score for a given base title"""
-        best_track = None
-        best_score = -1
-        for track in artist_tracks:
-            title = track.get("name", "").strip().lower()
-            track_base = self._extract_base_title(title)
-            if track_base == base_title:
-                score = track.get("similarity_score", 0)
-                if score > best_score:
-                    best_score = score
-                    best_track = track
-        return best_track
-
-    def _process_artist_tracks(
-        self, artist_tracks: List[Dict[str, Any]]
-    ) -> List[Dict[str, Any]]:
-        """Process tracks for a single artist to remove duplicates"""
-        if len(artist_tracks) == 1:
-            return [artist_tracks[0]]
-
-        deduplicated = []
-        processed_titles = set()
-
-        for track in artist_tracks:
-            title = track.get("name", "").strip().lower()
-            base_title = self._extract_base_title(title)
-
-            if base_title and base_title not in processed_titles:
-                processed_titles.add(base_title)
-                best_track = self._find_best_track_for_base_title(
-                    artist_tracks, base_title
-                )
-                if best_track:
-                    deduplicated.append(best_track)
-            elif not base_title:
-                deduplicated.append(track)
-
-        return deduplicated
-
-    def _smart_title_deduplication(
-        self, tracks: List[Dict[str, Any]]
-    ) -> List[Dict[str, Any]]:
-        """Smart deduplication for titles with slight variations (e.g., 'Song (Remix)' vs 'Song')"""
-        if len(tracks) <= 1:
-            return tracks
-
-        artist_groups = self._group_tracks_by_artist(tracks)
-        deduplicated = []
-
-        for artist_tracks in artist_groups.values():
-            deduplicated.extend(self._process_artist_tracks(artist_tracks))
-
-        return deduplicated
-
-    def _extract_base_title(self, title: str) -> str:
-        """Extract base title by removing common suffixes and variations"""
-        if not title:
-            return ""
-
-        # Remove common suffixes in parentheses
-        import re
-
-        base = re.sub(r"\s*\([^)]*\)\s*$", "", title)
-
-        # Remove common suffixes
-        suffixes_to_remove = [
-            " (remix)",
-            " (remastered)",
-            " (live)",
-            " (acoustic)",
-            " (radio edit)",
-            " (extended)",
-            " (clean)",
-            " (explicit)",
-            " (original mix)",
-            " (club mix)",
-            " (dub mix)",
-            " - remix",
-            " - remastered",
-            " - live",
-            " - acoustic",
-        ]
-
-        for suffix in suffixes_to_remove:
-            if base.lower().endswith(suffix.lower()):
-                base = base[: -len(suffix)]
-                break
-
-        return base.strip()
-
-    def _normalize_file_location(self, location: str) -> str:
-        """Normalize file location for deduplication and Apple Music compatibility"""
-        if not location:
-            return ""
-
-        # Convert to lowercase and normalize separators
-        normalized = location.lower().replace("\\", "/")
-
-        # Remove common OS-specific prefixes with username
-        import re
-
-        # Match /users/<username>/, /home/<username>/, c:/users/<username>/, etc.
-        match = re.match(
-            r"^(?:/users/|/home/|c:/users/|d:/users/|e:/users/)([^/]+)/(.+)$",
-            normalized,
-        )
-        if match:
-            return match.group(2)
-        return normalized
-
-    def save_playlist_m3u(
-        self, tracks: List[Dict[str, Any]], query: str, output_dir: str = "playlists"
-    ) -> str:
-        """
-        Save playlist as M3U file optimized for Apple Music
-
-        Args:
-            tracks: List of track dictionaries
-            query: Original search query
-            output_dir: Output directory for playlist files
-
-        Returns:
-            Path to saved playlist file
-        """
-        try:
-            # Create output directory if it doesn't exist
-            os.makedirs(output_dir, exist_ok=True)
-
-            # Create filename with timestamp and sanitized query
-            timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-            sanitized_query = re.sub(r"[^\w\s-]", "", query).strip()
-            sanitized_query = re.sub(r"[-\s]+", "-", sanitized_query)
-            sanitized_query = sanitized_query[:50]  # Limit length
-
-            filename = f"playlist_{timestamp}_{sanitized_query}.m3u"
-            filepath = os.path.join(output_dir, filename)
-
-            # Write M3U file optimized for Apple Music
-            with open(filepath, "w", encoding="utf-8") as f:
-                # Write M3U header
-                f.write("#EXTM3U\n")
-
-                # Write playlist metadata
-                f.write("# Generated by Tonal Hortator (Local)\n")
-                f.write(f"# Query: {query}\n")
-                f.write(
-                    f"# Generated: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}\n"
-                )
-                f.write(f"# Tracks: {len(tracks)}\n")
-                f.write("# Format: M3U Extended\n")
-                f.write("# Target: Apple Music\n\n")
-
-                # Write tracks
-                for i, track in enumerate(tracks, 1):
-                    # Get track info
-                    artist = track.get("artist", "Unknown")
-                    title = track.get("name", "Unknown")
-                    album = track.get("album", "Unknown")
-                    duration = track.get("duration_ms", 0) // 1000  # Convert to seconds
-                    similarity = track.get("similarity_score", 0)
-                    location = track.get("location", "")
-
-                    # Write extended info line
-                    # Format: #EXTINF:duration,artist - title
-                    f.write(f"#EXTINF:{duration},{artist} - {title}\n")
-
-                    # Write file path
-                    if location:
-                        # Convert file:// URLs to local paths for better Apple Music compatibility
-                        if location.startswith("file://"):
-                            # Remove file:// prefix and decode URL encoding
-                            import urllib.parse
-
-                            local_path = urllib.parse.unquote(location[7:])
-                            f.write(f"{local_path}\n")
-                        else:
-                            f.write(f"{location}\n")
-                    else:
-                        # If no location, write a comment
-                        f.write(f"# Missing file location for: {artist} - {title}\n")
-
-                    # Write additional metadata as comments
-                    f.write(f"# Album: {album}\n")
-                    f.write(f"# Similarity: {similarity:.3f}\n")
-                    f.write(f"# Track: {i}/{len(tracks)}\n\n")
-
-            logger.info(f"💾 Saved Apple Music playlist to: {filepath}")
-            return filepath
-
-        except Exception as e:
-            logger.error(f"❌ Error saving playlist: {e}")
-            raise
-
-    def print_playlist_summary(self, tracks: List[Dict[str, Any]], query: str) -> None:
-        """Print a summary of the generated playlist"""
-        print(f"\n🎵 Playlist for: '{query}'")
-        print(f"📊 {len(tracks)} tracks found")
-        print("-" * 60)
-
-        for i, track in enumerate(tracks, 1):
-            artist = track.get("artist", "Unknown")
-            name = track.get("name", "Unknown")
-            album = track.get("album", "Unknown")
-            similarity = track.get("similarity_score", 0)
-
-            print(f"{i:2d}. {artist} - {name}")
-            print(f"    Album: {album}")
-            print(f"    Similarity: {similarity:.3f}")
-            print()
-
-        if tracks:
-            avg_similarity = sum(t.get("similarity_score", 0) for t in tracks) / len(
-                tracks
-            )
-            print(f"📈 Average similarity: {avg_similarity:.3f}")
-        print("-" * 60)
-
-    def _extract_track_count(self, query: str) -> Optional[int]:
-        """Extract desired track count from query using regex"""
-        # Patterns to look for a number followed by "tracks" or "songs"
-        # e.g., "15 tracks", "a playlist of 10 songs", "top 20"
-        patterns = [
-            r"(\d+)\s+(?:tracks|songs)",
-            r"top\s+(\d+)",
-            r"playlist\s+of\s+(\d+)",
-        ]
-
-        for pattern in patterns:
-            match = re.search(pattern, query, re.IGNORECASE)
-            if match:
-                return int(match.group(1))
-
-        return None
-
-    def _is_vague_query(self, query: str) -> bool:
-        """
-        Detect if a query is vague and would benefit from artist randomization
-
-        Args:
-            query: Search query
-
-        Returns:
-            True if query is considered vague
-        """
-        query_lower = query.lower().strip()
-
-        # Vague query patterns that typically return many artists
-        vague_patterns = [
-            # Generic music types
-            r"\b(video game|game)\s+music\b",
-            r"\b(early|late|mid)\s+\d{2}s\s+\w+\b",  # e.g., "early 90s grunge"
-            r"\b(upbeat|happy|sad|melancholy|energetic|chill|relaxing)\s+(music|songs?|tracks?)\b",
-            r"\b(party|workout|study|sleep|morning|evening|night)\s+(music|songs?|playlist)\b",
-            r"\b(classic|old|vintage|retro|nostalgic)\s+(music|songs?)\b",
-            r"\b(modern|contemporary|new|recent)\s+(music|songs?)\b",
-            r"\b(background|ambient|atmospheric)\s+(music|songs?)\b",
-            r"\b(instrumental|acoustic|electronic|digital)\s+(music|songs?)\b",
-            # Generic genres without specific artists
-            r"\b(rock|pop|jazz|blues|country|folk|electronic|hip.?hop|rap|r&b|soul|funk|disco|punk|metal|indie|alternative)\s+(music|songs?)\b",
-            # Very short queries
-            r"^\w{1,3}$",  # Single words or very short phrases
-        ]
-
-        for pattern in vague_patterns:
-            if re.search(pattern, query_lower):
-                return True
-
-        # Check if query is very short (likely vague)
-        if len(query_lower.split()) <= 2:
-            return True
-
-        return False
-
-    def _apply_artist_randomization(
-        self, tracks: List[Dict[str, Any]], max_tracks: int, is_vague: bool
-    ) -> List[Dict[str, Any]]:
-        """
-        Apply artist randomization to increase diversity for vague queries
-
-        Args:
-            tracks: List of tracks to process
-            max_tracks: Maximum number of tracks to return
-            is_vague: Whether the query is considered vague
-
-        Returns:
-            List of tracks with improved artist diversity
-        """
-        if not is_vague or len(tracks) <= max_tracks:
-            return tracks[:max_tracks]
-
-        # Group tracks by artist
-        artist_groups: dict[str, list[dict[str, Any]]] = {}
-        for track in tracks:
-            artist = track.get("artist", "").strip()
-            if artist:
-                if artist not in artist_groups:
-                    artist_groups[artist] = []
-                artist_groups[artist].append(track)
-
-        # If we have many artists, prioritize diversity
-        if (
-            len(artist_groups) > max_tracks // 2
-        ):  # More artists than half the max tracks
-            logger.info(
-                f"🎤 Vague query detected. Randomizing {len(artist_groups)} artists for diversity"
-            )
-
-            # Sort artists by their best track's similarity score
-            artist_scores = []
-            for artist, artist_tracks in artist_groups.items():
-                best_score = max(
-                    track.get("similarity_score", 0) for track in artist_tracks
-                )
-                artist_scores.append((artist, best_score, artist_tracks))
-
-            # Sort by score (descending)
-            artist_scores.sort(key=lambda x: x[1], reverse=True)
-
-            # Take top artists but randomize their order slightly
-            top_artists = artist_scores[:max_tracks]
-
-            # Shuffle the order of artists (but keep top ones)
-            if len(top_artists) > 3:
-                # Keep top 3 artists in order, shuffle the rest
-                top_3 = top_artists[:3]
-                rest = top_artists[3:]
-                random.shuffle(rest)
-                top_artists = top_3 + rest
-
-            # Take best track from each artist
-            randomized_tracks = []
-            for artist, _, artist_tracks in top_artists:
-                # Sort tracks by similarity score and take the best one
-                best_track = max(
-                    artist_tracks, key=lambda t: t.get("similarity_score", 0)
-                )
-                randomized_tracks.append(best_track)
-
-                if len(randomized_tracks) >= max_tracks:
-                    break
-
-            return randomized_tracks
-        else:
-            # Not enough artists to randomize, return original order
-            return tracks[:max_tracks]
-
-    def _enforce_artist_diversity(
-        self,
-        tracks: List[Dict[str, Any]],
-        max_tracks: int,
-        max_tracks_per_artist: int = 3,
-    ) -> List[Dict[str, Any]]:
-        """
-        Enforce artist diversity by limiting tracks per artist
-
-        Args:
-            tracks: List of tracks to process
-            max_tracks: Maximum number of tracks to return
-            max_tracks_per_artist: Maximum number of tracks allowed per artist
-
-        Returns:
-            List of tracks with enforced artist diversity
-        """
-        if not tracks:
-            return tracks
-
-        # Group tracks by artist
-        artist_groups: dict[str, list[dict[str, Any]]] = {}
-        for track in tracks:
-            artist = track.get("artist", "").strip()
-            if artist:
-                if artist not in artist_groups:
-                    artist_groups[artist] = []
-                artist_groups[artist].append(track)
-            else:
-                # Handle tracks without artist info
-                if "unknown" not in artist_groups:
-                    artist_groups["unknown"] = []
-                artist_groups["unknown"].append(track)
-
-        logger.info(
-            f"🎤 Enforcing artist diversity: {len(artist_groups)} artists, max {max_tracks_per_artist} tracks per artist"
-        )
-
-        # Process each artist group
-        diverse_tracks = []
-        for artist, artist_tracks in artist_groups.items():
-            if len(artist_tracks) > max_tracks_per_artist:
-                # Sort by similarity score and take the best tracks
-                sorted_tracks = sorted(
-                    artist_tracks,
-                    key=lambda t: t.get("similarity_score", 0),
-                    reverse=True,
-                )
-                # Take only the best tracks up to the limit
-                limited_tracks = sorted_tracks[:max_tracks_per_artist]
-                diverse_tracks.extend(limited_tracks)
-                logger.info(
-                    f"🎵 Limited {artist}: {len(artist_tracks)} → {len(limited_tracks)} tracks"
-                )
-            else:
-                # Artist is within limit, keep all tracks
-                diverse_tracks.extend(artist_tracks)
-
-        # Sort by similarity score and limit to max_tracks
-        diverse_tracks.sort(key=lambda t: t.get("similarity_score", 0), reverse=True)
-        final_tracks = diverse_tracks[:max_tracks]
-
-        # Distribute artists throughout the playlist to avoid grouping
-        distributed_tracks = self._distribute_artists(final_tracks)
-
-        logger.info(
-            f"✅ Artist diversity enforcement: {len(tracks)} → {len(distributed_tracks)} tracks"
-        )
-        return distributed_tracks
-
-    def _distribute_artists(self, tracks: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
-        """
-        Distribute artists throughout the playlist to avoid grouping
-
-        Args:
-            tracks: List of tracks to distribute
-
-        Returns:
-            List of tracks with distributed artists
-        """
-        if len(tracks) <= 1:
-            return tracks
-
-        # Group tracks by artist
-        artist_groups = self._group_tracks_by_artist(tracks)
-
-        # If we only have one artist or very few tracks, no need to distribute
-        if len(artist_groups) <= 1 or len(tracks) <= 3:
-            return tracks
-
-        logger.info(
-            f"🎯 Distributing {len(artist_groups)} artists across {len(tracks)} tracks"
-        )
-
-        # Create a distribution pattern to spread artists evenly
-        distributed: List[Dict[str, Any]] = []
-        artist_queues = self._create_artist_queues(artist_groups)
-
-        # Distribute tracks using round-robin with randomization
-        distributed = self._round_robin_distribute(artist_queues, tracks, distributed)
-
-        # Add any remaining tracks
-        distributed = self._add_remaining_tracks(artist_queues, tracks, distributed)
-
-        logger.info(
-            f"🎵 Artist distribution complete: {len(tracks)} tracks distributed"
-        )
-        return distributed
-
-    def _create_artist_queues(
-        self, artist_groups: dict[str, list[dict[str, Any]]]
-    ) -> dict[str, list[dict[str, Any]]]:
-        """Create queues for each artist's tracks"""
-        artist_queues: dict[str, list[dict[str, Any]]] = {}
-        for artist, artist_tracks in artist_groups.items():
-            artist_queues[artist] = artist_tracks.copy()
-        return artist_queues
-
-    def _round_robin_distribute(
-        self,
-        artist_queues: dict[str, list[dict[str, Any]]],
-        tracks: List[Dict[str, Any]],
-        distributed: List[Dict[str, Any]],
-    ) -> List[Dict[str, Any]]:
-        """Distribute tracks using round-robin with randomization"""
-        while artist_queues and len(distributed) < len(tracks):
-            # Get all artists that still have tracks
-            available_artists = list(artist_queues.keys())
-
-            # Add some randomization to avoid predictable patterns
-            if len(available_artists) > 1:
-                # Shuffle the order occasionally to add variety
-                if len(distributed) % 3 == 0:  # Every 3rd track
-                    random.shuffle(available_artists)
-
-            # Take one track from each available artist in round-robin fashion
-            for artist in available_artists:
-                if len(distributed) >= len(tracks):
-                    break
-
-                if artist_queues[artist]:
-                    track = artist_queues[artist].pop(0)
-                    distributed.append(track)
-
-                    # Remove artist from queue if no more tracks
-                    if not artist_queues[artist]:
-                        del artist_queues[artist]
-
-        return distributed
-
-    def _add_remaining_tracks(
-        self,
-        artist_queues: dict[str, list[dict[str, Any]]],
-        tracks: List[Dict[str, Any]],
-        distributed: List[Dict[str, Any]],
-    ) -> List[Dict[str, Any]]:
-        """Add any remaining tracks to the distributed list"""
-        # If we still have space and tracks left, add remaining tracks
-        remaining_tracks: List[Dict[str, Any]] = []
-        for artist_tracks in artist_queues.values():
-            remaining_tracks.extend(artist_tracks)
-
-        # Add remaining tracks at the end, maintaining artist diversity
-        if remaining_tracks and len(distributed) < len(tracks):
-            # Sort remaining tracks by similarity and add them
-            remaining_tracks.sort(
-                key=lambda t: t.get("similarity_score", 0), reverse=True
-            )
-            space_left = len(tracks) - len(distributed)
-            distributed.extend(remaining_tracks[:space_left])
-
-        return distributed
-
-
-def _check_embeddings_available(generator: LocalPlaylistGenerator) -> bool:
-    """Check if embeddings are available in the database"""
-    stats = generator.track_embedder.get_embedding_stats()
-    logger.info(f"📊 Database stats: {stats}")
-
-    if stats["tracks_with_embeddings"] == 0:
-        logger.warning("No embeddings found. Please run embed_tracks_local.py first.")
-        return False
-    return True
-
-
-def _process_playlist_request(generator: LocalPlaylistGenerator, query: str) -> bool:
-    """Process a single playlist request"""
-    try:
-        # Generate playlist
-        tracks = generator.generate_playlist(query, max_tracks=20)
-
-        if not tracks:
-            print("❌ No tracks found for your query. Try a different search term.")
-            return False
-
-        # Print summary
-        generator.print_playlist_summary(tracks, query)
-
-        # Ask if user wants to save
-        save = input("\nSave playlist to file? (y/n): ").strip().lower()
-        if save in ["y", "yes"]:
-            filepath = generator.save_playlist_m3u(tracks, query)
-            print(f"✅ Playlist saved to: {filepath}")
-
-            # Ask if user wants to open in Apple Music
-            open_music = input("Open in Apple Music? (y/n): ").strip().lower()
-            if open_music in ["y", "yes"]:
-                try:
-                    import subprocess
-
-                    subprocess.run(["open", "-a", "Music", filepath], check=True)
-                    print("🎵 Opened in Apple Music!")
-                except Exception as e:
-                    print(f"❌ Could not open in Apple Music: {e}")
-                    print(
-                        "💡 You can open it manually with: python open_in_apple_music.py"
-                    )
-
-        return True
-
-    except Exception as e:
-        logger.error(f"❌ Error generating playlist: {e}")
-        print(f"❌ Error: {e}")
-        return False
-
-
-def _run_interactive_loop(generator: LocalPlaylistGenerator) -> None:
-    """Run the interactive playlist generation loop"""
-    print("\n🎵 Local Playlist Generator")
-    print("=" * 40)
-
-    while True:
-        query = input("\nEnter your playlist query (or 'quit' to exit): ").strip()
-
-        if query.lower() in ["quit", "exit", "q"]:
-            break
-
-        if not query:
-            print("Please enter a query.")
-            continue
-
-        _process_playlist_request(generator, query)
-
-    print("\n👋 Thanks for using the Local Playlist Generator!")
-
-
-def main() -> None:
-    """Main function for interactive playlist generation"""
-    try:
-        logger.info("🚀 Starting local playlist generator")
-
-        # Initialize generator
-        generator = LocalPlaylistGenerator()
-
-        # Check if embeddings are available
-        if not _check_embeddings_available(generator):
-            return
-
-        # Run interactive loop
-        _run_interactive_loop(generator)
-
-    except Exception as e:
-        logger.error(f"❌ Error in main function: {e}")
-        raise
-
-
-if __name__ == "__main__":
-    main()
